(BSR) refactor(ui): move EmailInput to his own folder
