+++ src/features/auth/forgottenPassword/ForgottenPassword/ForgottenPassword.tsx
@@ -16,7 +16,7 @@ import { BottomContentPage } from 'ui/components/BottomContentPage'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { Form } from 'ui/components/Form'
 import { isEmailValid } from 'ui/components/inputs/emailCheck'
-import { EmailInput } from 'ui/components/inputs/EmailInput'
+import { EmailInput } from 'ui/components/inputs/EmailInput/EmailInput'
 import { isValueEmpty } from 'ui/components/inputs/helpers'
 import { InputError } from 'ui/components/inputs/InputError'
 import { ModalHeader } from 'ui/components/modals/ModalHeader'
