+++ src/features/profile/pages/ChangeEmail/ChangeEmail.tsx
@@ -22,7 +22,7 @@ import { theme } from 'theme'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { Form } from 'ui/components/Form'
 import { PageHeaderSecondary } from 'ui/components/headers/PageHeaderSecondary'
-import { EmailInput } from 'ui/components/inputs/EmailInput'
+import { EmailInput } from 'ui/components/inputs/EmailInput/EmailInput'
 import { InputError } from 'ui/components/inputs/InputError'
 import { PasswordInput } from 'ui/components/inputs/PasswordInput'
 import { useForHeightKeyboardEvents } from 'ui/components/keyboard/useKeyboardEvents'
